[DSCP-126] Use Word32 for nonce everywhere
This is important fix, as soon as before this validation used `Integer` for
nonce but we want nonces set to form a round.

+++ witness/src/Dscp/Witness/Web/Logic.hs
@@ -72,7 +72,7 @@ toAccountInfo account txs = do
     pure AccountInfo
         { aiBalances = Coin . fromIntegral . aBalance <$> account
         , aiCurrentNonce = nonce
-        , aiTransactionCount = nonce
+        , aiTransactionCount = fromIntegral nonce
         , aiTransactions = transactions
         }
   where
